não faz sentido

+++ mandatory/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: igenial <igenial@student.42sp.org.br>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/29 13:45:23 by igenial           #+#    #+#             */
-/*   Updated: 2023/07/11 17:55:18 by igenial          ###   ########.fr       */
+/*   Updated: 2023/07/12 19:04:42 by igenial          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	ft_type(const char *print, va_list action_object);
 int	ft_putchar(char c);
 int	ft_putstr(char *s);
 int	ft_putnbr(long nb);
-int	ft_puthex(size_t nb, char x);
+int	ft_puthex(unsigned int nb, char x);
 int	ft_putptr(unsigned long nb);
 int	ft_printf(const char *str, ...);
